+++ Spoony-iOS/Spoony-iOS/Source/Feature/Home/Home.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import NMapsMap
 
 struct Home: View {
     
@@ -20,24 +19,6 @@ struct Home: View {
         }
     }
 }
-
-struct NMapView: UIViewRepresentable {
-    func makeUIView(context: Context) -> NMFMapView {
-        let mapView = NMFMapView()
-        mapView.positionMode = .direction
-        mapView.zoomLevel = 15
-        
-        //시청으로 임시 고정
-        let cameraUpdate = NMFCameraUpdate(scrollTo: NMGLatLng(lat: 37.5666102, lng: 126.9783881))
-        mapView.moveCamera(cameraUpdate)
-        
-        return mapView
-    }
-    
-    func updateUIView(_ uiView: NMFMapView, context: Context) {
-    }
-}
-
 #Preview {
     Home()
         .environmentObject(NavigationManager())
